backend/classroom-service-py/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic==2.8.2
python-jose[cryptography]
python-multipart